add initRepoReleaseNew
